refactor grid to a stateless component
